PackageManager: Always let UID 0 start FallbackHome
com.android.settings/.FallbackHome is started at the end of the boot
process, before the UI is setup. If com.android.settings is protected
FallbackHome can't start and it's not possible to manually give it the
permission to start. Fix this allowing UID 0 start FallbackHome even
if com.android.settings is protected.

BUGBASH-103

Change-Id: I92bab937c387f9360de833be6779198a508914fe