base: Expand header image size
This shows the omnistyle header on a bigger size, taking all the statusbar instead of leaving the QS Footer clean.
Made it like old nougat times

Also, taken the way that omni made the padding between the qs header and the icon when this is enabled and recreate it just for this.

Change-Id: Ia74dbd81df6b47f4ffb50039116bc2a10323b46b