NotificationManagerService: do not use flashing API for staying always on
We allow to let the user set, that the notification light should stay
always on instead of flashing.

For that we are setting on duration to 1 and off duration to 0.
This can confuse some light HALs and/or kernel drivers.

Instead of using setFlashing, use setColor if duration is 1:0, like
we are already using for battery lights.

Change-Id: I4953592c72e0d095178757211dce2e206d8adb3c
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>